Move TemporarySoftHSM to __init__.py
